# Conflicts:
#   Frontend/VolleyTesting/app/src/main/java/com/example/volleytesting/Friends.java
#   Frontend/VolleyTesting/app/src/main/res/layout/activity_user_list.xml

+++ Backend/309_Server/src/main/java/callyourmom/controller/EventController.java
@@ -20,7 +20,7 @@ public class EventController {
 	}
 	
 	@GetMapping("/events/{id}")
-	Event get(@PathVariable Long id) {
+	Event get(@PathVariable Integer id) {
 	    return eventService.get(id);
 	}
